Refactor PackageConfigGetter
Move the class to a separate file as it isn't directly related to service config and to avoid circular imports.
Move the create_issue_if_needed to the reporters module.
