+++ packit_service/events/pagure/pr.py
@@ -9,7 +9,8 @@ from ogr.abstract import GitProject
 from ogr.parsing import RepoUrl
 from packit.config import PackageConfig
 
-from packit_service.config import PackageConfigGetter, ServiceConfig
+from packit_service.config import ServiceConfig
+from packit_service.package_config_getter import PackageConfigGetter
 from packit_service.service.db_project_events import (
     AddPullRequestEventToDb,
 )
