+++ packit_service/worker/reporting/__init__.py
@@ -9,6 +9,7 @@ from packit_service.worker.reporting.reporters.github import (
 )
 from packit_service.worker.reporting.reporters.gitlab import StatusReporterGitlab
 from packit_service.worker.reporting.utils import (
+    create_issue_if_needed,
     report_in_issue_repository,
     update_message_with_configured_failure_comment_message,
 )
@@ -22,4 +23,5 @@ __all__ = [
     StatusReporterGithubChecks.__name__,
     StatusReporterGithubStatuses.__name__,
     StatusReporterGitlab.__name__,
+    create_issue_if_needed.__name__,
 ]
