+++ tests/integration/test_issue_comment.py
@@ -20,7 +20,7 @@ from packit.exceptions import PackitException
 from packit.local_project import LocalProject, LocalProjectBuilder
 from packit.utils.koji_helper import KojiHelper
 
-from packit_service.config import PackageConfigGetter, ServiceConfig
+from packit_service.config import ServiceConfig
 from packit_service.constants import COMMENT_REACTION, TASK_ACCEPTED
 from packit_service.events import github, gitlab
 from packit_service.models import (
@@ -39,6 +39,7 @@ from packit_service.models import (
     SyncReleaseTargetModel,
     SyncReleaseTargetStatus,
 )
+from packit_service.package_config_getter import PackageConfigGetter
 from packit_service.service.urls import get_propose_downstream_info_url
 from packit_service.worker.allowlist import Allowlist
 from packit_service.worker.celery_task import CeleryTask
