+++ tests/unit/events/test_copr.py
@@ -8,10 +8,10 @@ import pytest
 from flexmock import flexmock
 from ogr.services.github import GithubProject
 
-from packit_service.config import PackageConfigGetter
 from packit_service.events.copr import End, Start
 from packit_service.events.enums import FedmsgTopic
 from packit_service.models import CoprBuildTargetModel, get_most_recent_targets
+from packit_service.package_config_getter import PackageConfigGetter
 from packit_service.worker.parser import Parser
 from tests.spellbook import DATA_DIR
