+++ tests/unit/events/test_gitlab.py
@@ -7,7 +7,6 @@ import pytest
 from flexmock import flexmock
 from ogr.services.gitlab import GitlabProject
 
-from packit_service.config import PackageConfigGetter
 from packit_service.events import abstract
 from packit_service.events.gitlab import (
     enums,
@@ -18,6 +17,7 @@ from packit_service.events.gitlab import (
     release,
 )
 from packit_service.models import PullRequestModel
+from packit_service.package_config_getter import PackageConfigGetter
 from packit_service.worker.parser import Parser
 from tests.spellbook import DATA_DIR
