+++ tests/unit/events/test_pagure.py
@@ -7,12 +7,12 @@ import pytest
 from flexmock import flexmock
 from ogr.services.pagure import PagureProject
 
-from packit_service.config import PackageConfigGetter
 from packit_service.events.enums import PullRequestAction, PullRequestCommentAction
 from packit_service.events.pagure import (
     pr,
     push,
 )
+from packit_service.package_config_getter import PackageConfigGetter
 from packit_service.worker.parser import Parser
 from tests.spellbook import DATA_DIR
